03-implementation/requirements.txt
# TTS Project Dependencies
edge-tts>=6.1.0
ffmpeg-python>=0.2.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
